clippy

+++ pil2-components/lib/std/rs/src/std_virtual_table.rs
@@ -216,8 +216,7 @@ impl VirtualTableAir {
             let row_idx = offset & self.mask;
 
             // Update the multiplicity
-            self.multiplicities[sub_table_idx as usize][row_idx as usize]
-                .fetch_add(multiplicity as u64, Ordering::Relaxed);
+            self.multiplicities[sub_table_idx as usize][row_idx as usize].fetch_add(multiplicity, Ordering::Relaxed);
         }
     }
 }
